% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{add_bin_labels}
\alias{add_bin_labels}
\title{Reverse minmax scaling of numeric vector}
\usage{
add_bin_labels(x, one, zero)
}
\arguments{
\item{x}{a numeric vector or column, scaled between 0 and 1.}

\item{one}{a character string, the label associated with binary value 1}

\item{zero}{a character string, the label associated with binary value 0}
}
\description{
Helper function to re-apply binary variable labels post-imputation.
}
\examples{
ex_bin <- --c(1,0,0,1,1,0,0,1,0)
cat <- "cat"
dog <- "dog"

add_bin_labels(x = ex_bin, one = cat, zero = dog)
}
\keyword{postprocessing}
