% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_imputation.R
\name{coalesce_one_hot}
\alias{coalesce_one_hot}
\title{Coalesce one-hot encoding back to a single variable}
\usage{
coalesce_one_hot(X, var_name)
}
\arguments{
\item{X}{a data.frame, data.table or matrix, for a single variable.}

\item{cat_names}{a character vector, containing one-hot column names}
}
\description{
Helper function to reverse one-hot encoding post-imputation.
}
\examples{
ex_bin <- --c(1,0,0,1,1,0,0,1,0)
cat <- "cat"
dog <- "dog"

add_bin_labels(x = ex_bin, one = cat, zero = dog)
}
\keyword{postprocessing}
