% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{col_minmax}
\alias{col_minmax}
\title{Scales numeric vector between 0 and 1}
\usage{
col_minmax(x)
}
\arguments{
\item{x}{a numeric vector or column.}
}
\description{
Helper function to scale numeric variables. Aids convergence of Midas model.
}
\examples{
ex_num <- runif(100,1,10)
scaled <- col_minmax(ex_num)
}
\keyword{preprocessing}
