% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rubin_analysis.R
\name{combine}
\alias{combine}
\title{Calculate and combine \emph{m} regression models}
\usage{
combine(formula, df_list, dof_adjust = TRUE, ...)
}
\arguments{
\item{formula}{a model formula or string coercible to a formula}

\item{df_list}{a list of completed data.frames or objects coercible to data.frames}

\item{dof_adjust}{a boolean, indicating whether or not to apply the Rubin and Barnard (1999) degrees of freedom adjustment for small-samples}

\item{...}{Further arguments passed onto \code{glm()}, enabling users to run other model families.}
}
\value{
Dataframe of combined model results.
}
\description{
\code{combine()} calculates \emph{m} individual regression models, then applies "Rubin's Rules" (1987) to produce a single, combined estimate of the regression parameters and uncertainty.
}
\examples{
# Generate raw data, with numeric, binary, and categorical variables
raw_data = data.frame(a = sample(c("red","yellow","blue",NA),1000, replace = TRUE),
                      b = 1:1000,
                      c = sample(c("YES","NO",NA),1000,replace=TRUE),
                      d = runif(1000,1,10),
                      e = sample(c("YES","NO"), 1000, replace = TRUE),
                      f = sample(c("male","female","trans","other",NA), 1000, replace = TRUE))

# Names of bin./cat. variables
test_bin <- c("c","e")
test_cat <- c("a","f")

# Pre-process data
test_data <- convert(raw_data,
                     bin_cols = test_bin,
                     cat_cols = test_cat,
                     minmax_scale = TRUE)

# Train imputation model
test_model <- train(test_data)

# Generate datasets
complete_datasets <- complete(test_imp, m = 5)

combine(formula = d~a+c+e+f, complete_datasets)
}
\keyword{import}
