% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_env.R
\name{delete_rMIDAS_env}
\alias{delete_rMIDAS_env}
\title{Delete the rMIDAS Environment and Configuration}
\usage{
delete_rMIDAS_env()
}
\value{
A message indicating the completion of the deletion process.
}
\description{
Deletes both the virtual environment and the configuration file for the rMIDAS package.
After deletion, it is necessary to restart the R session and then load the rMIDAS package once more.
This will trigger the setup process again.
}
