% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midas_functions.R
\name{generate}
\alias{generate}
\title{Generate}
\usage{
generate(mid_obj, m = 10, file = NULL)
}
\description{
Having generated an imputation model, generate() produces \code{m} datasets, saved as a list.
}
\examples{
midas_obj <- impute(example_data,
                    layer_structure = c(128,128),
                    input_drop = 0.75,
                    learn_rate = 0.0005,
                    seed = 89)

imp_data <- generate(midas_obj,
                     m = 5)

}
\keyword{generate}
