% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{mid_py_setup}
\alias{mid_py_setup}
\title{Configure python for MIDAS imputation}
\usage{
mid_py_setup()
}
\description{
This helper function checks if the required Python dependencies are installed, and if not, checks with user before installing them.
Users should not call this function directly. Users can set a custom python install using \code{set_python_env()} so long as this is done prior to the first call to \code{train()} or \code{complete()}.
}
\keyword{setup}
