% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{py_setup}
\alias{py_setup}
\title{Configure python for MIDAS imputation}
\usage{
py_setup()
}
\description{
This function checks if the required Python dependencies are installed, and if not, checks with user before installing them.
This function is called automatically within \code{set_python_env}, and should only need to be used when manually configuring python installs using reticulate.
}
\keyword{setup}
