% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_utils.R
\name{python_configured}
\alias{python_configured}
\title{Check whether Python is capable of executing example code}
\usage{
python_configured()
}
\value{
\code{NULL}
}
\description{
Checks if each Python dependency is available.
This function is called within some examples to ensure code executes properly.
}
\keyword{setup}
