% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{python_init}
\alias{python_init}
\title{Initialise connection to Python}
\usage{
python_init()
}
\description{
Internal function. Checks if Python has already been initialised, and if not, completes the required setup to run the MIDAS algorithm.
This function is called automatically, and users should not call it directly.
To configure which Python install/environment/conda is used, see documentation for \code{set_python_env()}.
}
\keyword{setup}
