% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMIDAS-package.R
\docType{package}
\name{rMIDAS-package}
\alias{rMIDAS}
\alias{rMIDAS-package}
\title{rMIDAS: Multiple Imputation with Denoising Autoencoders}
\description{
\code{rMIDAS} provides an R interface to the MIDAS multiple imputation
algorithm.
}
\details{
\code{rMIDAS} is deprecated in favor of
\href{https://CRAN.R-project.org/package=rMIDAS2}{\code{rMIDAS2}},
which provides the actively maintained successor interface and backend.
The source repository for the successor package is
\url{https://github.com/MIDASverse/rMIDAS2}. \code{rMIDAS}
remains available on CRAN for existing workflows but will receive only
compatibility and documentation updates.

See \code{vignette("migrating-to-rMIDAS2", package = "rMIDAS")} for guidance
on moving to \code{rMIDAS2}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MIDASverse/rMIDAS}
  \item \url{https://CRAN.R-project.org/package=rMIDAS2}
  \item \url{https://github.com/MIDASverse/rMIDAS2}
  \item Report bugs at \url{https://github.com/MIDASverse/rMIDAS/issues}
}

}
\author{
\strong{Maintainer}: Thomas Robinson \email{ts.robinson1994@gmail.com} (\href{https://orcid.org/0000-0001-7097-1599}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Thomas Robinson \email{ts.robinson1994@gmail.com} (\href{https://orcid.org/0000-0001-7097-1599}{ORCID}) [copyright holder]
  \item Ranjit Lall \email{R.Lall@lse.ac.uk} (\href{https://orcid.org/0000-0003-1455-3506}{ORCID}) [copyright holder]
}

Other contributors:
\itemize{
  \item Alex Stenlake [contributor, copyright holder]
  \item Elviss Dvinskis [contributor]
}

}
\keyword{package}
