% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{set_python_env}
\alias{set_python_env}
\title{Configure python version}
\usage{
set_python_env(path = "~/.path/to/python/binary")
}
\arguments{
\item{path}{Path to python binary if \code{type == "auto"}, path to virtualenv if \code{type == "virtualenv"}, or the name of a Conda environment if \verb{type=="condaenv}}

\item{type}{One of 'auto','virtualenv','condaenv'.}

\item{exact}{Boolean. If \code{TRUE} then only load exact match from \code{path}, otherwise allow reticulate to scan for other versions.}
}
\description{
This function allows users to set a custom python binary, virtualenv, or Conda environment.
}
\keyword{setup}
