% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_utils.R
\name{skip_if_no_numpy}
\alias{skip_if_no_numpy}
\title{Skip test where 'numpy' not available.}
\usage{
skip_if_no_numpy()
}
\value{
\code{NULL}
}
\description{
Check if Python's numpy is available, and skip test if not.
This function is called within some tests to ensure server tests involving \code{reticulate} calls execute properly.
}
\keyword{setup}
