% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{skip_if_no_scipy}
\alias{skip_if_no_scipy}
\title{Another test function}
\usage{
skip_if_no_scipy()
}
\description{
This function cannot be used until i work out how testhat works
}
\examples{
skip_if_no_scipy()
}
\keyword{test}
