% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{test_function}
\alias{test_function}
\title{A test function}
\usage{
test_function(test_string)
}
\arguments{
\item{test_string}{Vector to be printed}
}
\description{
This function prints a test string.
}
\examples{
test_function()
}
\keyword{test}
