% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{coalesce_one_hot}
\alias{coalesce_one_hot}
\title{Coalesce one-hot encoding back to a single variable}
\usage{
coalesce_one_hot(X, var_name)
}
\arguments{
\item{X}{A data.frame, data.table or matrix, for a single variable}

\item{var_name}{A character string, with the original variable label}
}
\description{
Helper function to reverse one-hot encoding post-imputation.
}
\keyword{postprocessing}
