% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midas_functions.R
\name{complete}
\alias{complete}
\title{Impute missing values using imputation model}
\usage{
complete(mid_obj, m = 10L, file = NULL, file_root = NULL)
}
\arguments{
\item{mid_obj}{object of class \code{midas}, the result of running \code{rMIDAS::impute()}}

\item{m}{integer number of completed datasets required}

\item{file}{path to save completed datasets. If \code{NULL}, completed datasets are only loaded into memory.}

\item{file_root}{character vector used as root of completed datasets if a filepath is passed to function. If no file_root is provided, saved datasets will be saved as "file/midas_impute_<yymmdd_hhmmss>_\link{m}.csv"}
}
\description{
Having trained an imputation model, complete() produces \code{m} completed datasets, saved as a list.
}
\examples{
midas_obj <- train(example_data,
                    layer_structure = c(128,128),
                    input_drop = 0.75,
                    learn_rate = 0.0005,
                    seed = 89)

imp_data <- complete(midas_obj,
                     m = 5)

}
\keyword{impute}
