% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midas_functions.R
\name{import_midas}
\alias{import_midas}
\title{Instantiate Midas class}
\usage{
import_midas(...)
}
\arguments{
\item{...}{Arguments passed to the MIDAS class for instantiating network}
}
\description{
Import Midas class into R environment, and instantiates passed parameters.
}
\examples{
midas_base <- import_midas()
}
\keyword{import}
