% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{na_to_nan}
\alias{na_to_nan}
\title{Replace missing values}
\usage{
na_to_nan(df)
}
\arguments{
\item{df}{Data.frame or coercible}
}
\description{
Helper function to convert \code{NA} values in a data.frame to \code{NaN}. This ensures the correct conversion of missing values when reticulate converts R objects to their Python equivalent. See the reticulate package documentation on type conversions for more information.
}
\examples{
\dontrun{
na_conv <- na_to_nan(data.frame(a = c(1,NA,0,0,NA,NA)))
}
}
\keyword{preprocessing}
